% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{add_data_to_shapes}
\alias{add_data_to_shapes}
\title{Add data to shapes}
\usage{
add_data_to_shapes(shapes, data, by)
}
\arguments{
\item{shapes}{A SpatialPolygonsDataFrame as generated using the SP package}

\item{data}{The geospatial data that should be added to the shapes}

\item{by}{The column on the existing dataframe that new data should be joined to}
}
\value{
Returns the modified shape SpatialPolygonsDataFrame
}
\description{
Add data to shapes
}
\author{
Simon Raper
}

