% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{read_in_geouk}
\alias{read_in_geouk}
\title{Reads in and validates data for geoukr}
\usage{
read_in_geouk(file, sep)
}
\arguments{
\item{file}{The path and filename of a csv file}

\item{sep}{The column separator}
}
\value{
Returns a data frame, with validated geo locations and dates converted into the R date class
}
\description{
Reads in and validates data for geoukr
}
\author{
Simon Raper
}

