% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{ward_map_by_region}
\alias{ward_map_by_region}
\title{Plots a map for just one region of a uk ward map object}
\usage{
ward_map_by_region(region, uk_ward_map, breaks, map_title = NULL)
}
\arguments{
\item{region}{The name of the region. Possible values are: London, North West, Yorkshire and The Humber, North East, West Midlands, East Midlands, South West, East of England, South East}

\item{uk_ward_map}{A UK ward map created with the uk_ward_map function}

\item{map_title}{Optional title for the map}
}
\description{
Plots a map for just one region of a uk ward map object
}
\examples{
data(household_deprivation_3d)
dep_map<-uk_ward_map(household_deprivation_3d)
ward_map_by_region("London", dep_map)
}
\author{
Simon Raper
}

