% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{uk_ward_map}
\alias{uk_ward_map}
\title{Creates the data for a UK ward map by appending a ward level dataset to the ward shapes}
\usage{
uk_ward_map(ward_data, points = NULL)
}
\arguments{
\item{ward_data}{A ward level data set. This must contain a column called cm_ward_code that contains the ONS census merged ward codes and a column called value which contains the statistic to be mapped.}

\item{points}{An optional data frame containing the co-ordinates and labels of points that you may want to plot on the map}
}
\value{
Returns an object of class uk_ward_map containing the ward_shapes with the data appended and the la_shapes
}
\description{
Creates the data for a UK ward map by appending a ward level dataset to the ward shapes
}
\examples{
data(household_deprivation_3d)
dep_map<-uk_ward_map(household_deprivation_3d)
}
\author{
Simon Raper
}

